Hashmap fix. Compile warnings fix.

+++ classes/Component.cpp
@@ -12,7 +12,7 @@
 
 Component::Component(long UID, const char * name){
     this->UID = UID;
-    unsigned int name_length = std::strlen(name) + 1;
+    unsigned int name_length = (int)std::strlen(name) + 1;
     this->name = new char[name_length];
     std::memcpy(this->name, name, name_length);
 }
